Readability tweaks
Previously on the values section the text would be full width, making it
unreadable on large monitors. Similarly, titles would be far away from
the text they were referencing on large screens and all sections would
be a full screen height, regardless of the content, leading to a large
amount of blank space that had to be scrolled through.

Signed-off-by: Sam Whited <sam@samwhited.com>