Make single pages a single section
This removes the need for wrapping and re-rendering markdown in single
pages by adding the section we're currently using to the template. It
will also make updating it easier later since it only has to be done in
one place and not across all pages (assuming we end up with more of
those).

Signed-off-by: Sam Whited <sam@samwhited.com>