Rework home page to not require shortcodes
Instead of using shortcodes to build the home page, use a page bundle
and the template. This gets rid of all the confusing markdown vs. HTML
stuff and further separates the actual content from the layout.

This also required redoing the columns section to use grid instead of
flexbox as it ends up removing a div that flexbox would rely on.
Using grid means that as a future enhancement we could also get rid of
the "div" shortcode by putting titles and paragraphs into their own grid
areas.

Signed-off-by: Sam Whited <sam@samwhited.com>